# Optimize and enable aggregation for ugni