@ronawho 2019-12-09
Last commit is the important one, I'll rebase after https://github.com/mhmerrill/arkouda/pull/207 is merged.